+++ inlang/source-code/cross-sell/cross-sell-sherlock/src/recommendation/utils/isExtensionJson.ts
